Cache KC installation in CI

+++ install-kc-ci.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
-cd /opt
-curl -LO https://dbmx.net/kyotocabinet/pkg/kyotocabinet-1.2.79.tar.gz
-tar -xvzf kyotocabinet-1.2.79.tar.gz && mv kyotocabinet-1.2.79 kyotocabinet && rm kyotocabinet-1.2.79.tar.gz
+if [ ! -d "/opt/kyotocabinet" ]; then
+  cd /opt
+  curl -LO https://dbmx.net/kyotocabinet/pkg/kyotocabinet-1.2.79.tar.gz
+  tar -xvzf kyotocabinet-1.2.79.tar.gz && mv kyotocabinet-1.2.79 kyotocabinet && rm kyotocabinet-1.2.79.tar.gz
 
-apt-get update
-apt-get -y install liblzo2-dev liblzma-dev zlib1g-dev build-essential
-cd /opt/kyotocabinet && ./configure --enable-lzo --enable-lzma && make && make install
+  apt-get update
+  apt-get -y install liblzo2-dev liblzma-dev zlib1g-dev build-essential
+  cd /opt/kyotocabinet && ./configure --enable-lzo --enable-lzma && make && make install
+fi
